requirements.txt
flask
requests
groq
python-dotenv
psycopg2-binary
streamlit
pytest
pandas
scikit-learn